Optimize ExampleStatusParser
This commit replaces the example status parsing with a faster
implementation.

Instead of splitting each row by pipes, and zipping that with the
headers, we store the length and position of the each header (including
padding) and use that to extract the values.

This requires the assumption that the file being read is well formatted
and has the correct indentation on all lines.