Optimize Example.parse_id
This commit improves the performance of Example.parse_id at the expense
of making it a bit less readable. I've included the original in a
comment for this reason.

The existing regex was slow. I would guess partly because of the lazy
match. Calling .captures also seems to be slower than being explicit
about which captures to use.

This commit instead uses a regex to just match the final "scoped id"
within the brackets. It uses ruby's "special variables" to extract $`,
the content before the previous regex match, and $`, the first group of
the previous regex match.