Optimize sort_value_from
This commit switches to  String#to_i instead of Integer(). to_i is a
fair bit faster than integer.

    split.map(&method(:Integer))  557.625k (± 5.9%) i/s
    split.map { |x| Integer(x) }  881.760k (± 3.7%) i/s
             split.map(&:to_i)    937.586k (± 4.3%) i/s
            split.map!(&:to_i)    970.409k (± 2.8%) i/s

The differences between the two is that Integer will parse for prefixes
(like 0x) and will raise on errors. We are guaranteed to only get valid
base-10 integers here since we received them through parse_id (and we
would have output them originally anyways).

This switches map to map!. We're getting a new array from the call to
split(), and we might as well reuse it.

Finally, we use unshift instead of splatting the array, since it is
a little bit faster and removes an array allocation. The numbers here
are within the +-, but it does seem reliably better than the others.

          [file, *map]    888.896k (± 4.5%) i/s
     map.unshift(file)    976.050k (± 4.0%) i/s
    [file].concat(map)    927.339k (± 4.3%) i/s
          [file] + map    932.178k (± 4.0%) i/s